mv into SqlAst package
