+++ src/SqlAst/ConcatReturnTypeExtension.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace staabm\PHPStanDba\ParserExtension;
+namespace staabm\PHPStanDba\SqlAst;
 
 use PHPStan\Type\Accessory\AccessoryNonEmptyStringType;
 use PHPStan\Type\Accessory\AccessoryNonFalsyStringType;
